review: move aligned_allocator into /range/container
