+++ test/snippet/range/container/aligned_allocator.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <vector>
 
-#include <seqan3/memory/aligned_allocator.hpp>
+#include <seqan3/range/container/aligned_allocator.hpp>
 
 size_t memory_alignment(void * value, size_t alignment)
 {
