--- test/unit/memory/CMakeLists.txt
@@ -1,1 +1,0 @@
-seqan3_test(aligned_allocator_test.cpp)
