+++ test/unit/range/container/CMakeLists.txt
@@ -1,3 +1,4 @@
+seqan3_test(aligned_allocator_test.cpp)
 seqan3_test(container_concept_test.cpp)
 seqan3_test(container_of_container_test.cpp)
 seqan3_test(container_test.cpp)
